requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# AWS
boto3==1.34.0
botocore==1.34.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# AI/ML for abbreviation expansion
google-generativeai==0.3.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
structlog==23.2.0
